Drive_God_lin.c and Makefile Modified
First, the drive c file and the makefile were modified to be compatible
with windows along with linux. Second, the BPM status function has been
changed to always return success after performing calculations.  The
rejection of BPMs has been consilidated into the rest of Beta-Beat.
Finally, some minor changes to the code were made, see the comments at
the start of the c code for further details.

I apologize if this was redundant.