Back to 2J=sqrt(2J)^2, and Calculation fixed
For the purposes of testing the code, we have returned to the old way of
calculating 2J, but all the new calculations are kept and can be
implented.  There was a bug in finding the commonbpms in gen_kick_calc()
but this has been fixed due to new helper function in utilities.